rename factory file
