+++ backend/tournesol/tests/test_api_video_rate_later.py
@@ -6,7 +6,7 @@ from rest_framework.test import APIClient
 from core.models import User
 from core.tests.factories.user import UserFactory
 from tournesol.models import Entity, VideoRateLater
-from tournesol.tests.factories.video import VideoFactory, VideoRateLaterFactory
+from tournesol.tests.factories.entity import VideoFactory, VideoRateLaterFactory
 
 
 class VideoRateLaterApi(TestCase):
